+++ features/stamp_data.gd
@@ -2,4 +2,3 @@ class_name StampData extends Resource
 
 @export var texture: Texture2D
 @export var color: Color = Color.WHITE
-